A batch of updated artefact weapon tiles (roctavian, #11340)
For random artefacts and the unrands Majin-Bo, Wucad Mu, Leech, and
Wyrmbane. The latter is suppose to have its tile upgraded every couple
of enchantment, but I haven't yet made the code modifications for that.
For now, I've added the later variants and let the first variant be the
only one used.

Slings and fustibali have their ordinary and ego tiles updated in
addition to the artefact ones.