Updated entropy weaver and tormentor tiles (roctavian, #11340)
Entropy weavers don't have a multi-turn spell any more, so we don't need
3 tiles, but it was agreed that the second state is the best rendition
for the main tile.

The new tormentor is purple, which matches the console color better, but
it is *very* purple. Some would claim it is too purple, and I would have
no defense against this argument. Roctavian has said he'll redo the tile
entirely, but until then we'll use this one. The new tile helps make
tormentors distinct from, say, cacodemons, and telling demons apart is
the key to dominating the late-game, the end-game, extended, the
post-game, and the metagame.